requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord.py==1.7.3
discord==1.7.3
django-request==1.6.1
Django==4.0.5
idna==3.3
multidict==6.0.2
pip==22.1.2
prompt-toolkit==3.0.30
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
wheel==0.37.1
yarl==1.7.2
